src/requirements.txt
python-dotenv
requests~=2.24.0